% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multiply_numbers}
\alias{multiply_numbers}
\title{Multiply two numbers together}
\usage{
multiply_numbers(x, y)
}
\arguments{
\item{x}{a number.}

\item{y}{another number.}
}
\value{
The product of \code{x} and \code{y}.
}
\description{
Multiply two numbers together
}

